test_report_tool/requirements.txt
TestLink-API-Python-client~=0.8.1
pandas~=1.3.5
requests~=2.28.2
urllib3~=1.26.14